requirements.txt
# Core web / scraping
requests
beautifulsoup4

# RAG / embeddings / ML
sentence-transformers==3.1.1
transformers<=4.39.3
faiss-cpu
openai

# Data / utils
numpy==1.23.5  # safe with this torch/thinc stack
scikit-learn
scipy
tqdm
python-dotenv
Pillow

# PDF handling
PyMuPDF

# Testing
pytest

# Streamlit App
streamlit

# Spacy NLP stack (safe versions)
spacy>=3.7.2,<3.8.0
blis==0.7.9

# Explicit spacy model wheel install — Streamlit Cloud supports this syntax
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
